outlined refactor, began refactoring.
