+++ rpsls_main.py
@@ -1,0 +1,3 @@
+from rpsls.rpsls import main
+
+main()
